Move the StripWhiteSpace* iterators to the line_processor module
Also get rid of the convenience trait for creating those iterators. The
`LineIteratorExt` already fullfills that role.
